added initial docker files
