+++ docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3'
+
+services:
+  ipfs-endpoint:
+    image: ipfs-endpoint
+    build: .
+    environment:
+      NODE_ENV: production
+    ports:
+      - 8082:8082
+      - 8081:8081
+      - 5001:5001
+    links: go-ipfs
+  
+  go-ipfs:
+    image: jbenet/go-ipfs
