+++ helpers/key-hash.map.js
@@ -1,0 +1,14 @@
+let KeyHashMap = {
+
+    instance: null,
+
+    getOne: function() {
+        if (!this.instance && this.instance == undefined) {
+            this.instance = new Map();
+        }
+        return this.instance;
+    }
+
+};
+
+module.exports = KeyHashMap;
